+++ test/unit/encoder_test.exs
@@ -3,7 +3,7 @@ defmodule JSON.LD.EncoderTest do
 
   doctest JSON.LD.Encoder
 
-  alias RDF.{Dataset}
+  alias RDF.{Dataset, Graph, Description}
   alias RDF.NS
   alias RDF.NS.{XSD, RDFS}
 
@@ -19,7 +19,7 @@ defmodule JSON.LD.EncoderTest do
 
 
   def gets_serialized_to(input, output, opts \\ []) do
-    data_structs = Keyword.get(opts, :only, [Dataset])
+    data_structs = Keyword.get(opts, :data_structs, [Dataset, Graph])
     Enum.each data_structs, fn data_struct ->
       assert JSON.LD.Encoder.from_rdf!(data_struct.new(input), opts) == output
     end
@@ -36,7 +36,7 @@ defmodule JSON.LD.EncoderTest do
       |> gets_serialized_to([%{
             "@id"         => "http://a/b",
             "http://a/c"  => [%{"@id" => "http://a/d"}]
-          }])
+          }], data_structs: [Dataset, Graph, Description])
     end
 
     test "should generate object list" do
@@ -47,7 +47,7 @@ defmodule JSON.LD.EncoderTest do
               %{"@id" => "http://example.com/d"},
               %{"@id" => "http://example.com/e"}
             ]
-          }])
+          }], data_structs: [Dataset, Graph, Description])
     end
 
     test "should generate property list" do
@@ -56,7 +56,7 @@ defmodule JSON.LD.EncoderTest do
             "@id"                   => "http://example.com/b",
             "http://example.com/c"  => [%{"@id" => "http://example.com/d"}],
             "http://example.com/e"  => [%{"@id" => "http://example.com/f"}]
-          }])
+          }], data_structs: [Dataset, Graph, Description])
     end
 
     test "serializes multiple subjects" do
@@ -77,7 +77,7 @@ defmodule JSON.LD.EncoderTest do
       |> gets_serialized_to([%{
             "@id"                   => "http://example.com/a",
             "http://example.com/b"  => [%{"@value" => "foo", "@type" => "http://example.com/d"}]
-          }])
+          }], data_structs: [Dataset, Graph, Description])
     end
 
     test "integer" do
@@ -183,7 +183,7 @@ defmodule JSON.LD.EncoderTest do
       |> gets_serialized_to([%{
             "@id"                   => "_:a",
             "http://example.com/a"  => [%{"@id" => "http://example.com/b"}]
-          }])
+          }], data_structs: [Dataset, Graph, Description])
     end
 
     test "should generate blank nodes as object" do
@@ -409,7 +409,7 @@ defmodule JSON.LD.EncoderTest do
     |> Enum.each(fn ({title, data}) ->
          @tag data: data
          test title, %{data: %{input: input, output: output}} do
-            input |> gets_serialized_to(output, only: [Dataset])
+            input |> gets_serialized_to(output, data_structs: [Dataset])
          end
        end)
   end